requirements.txt
flask
gunicorn
pybase64
Authlib
google-cloud-logging
flask-cors
swagger-gen
protobuf
oauth2client